ref: token middleware

+++ part4/blog/controllers/blogs.js
@@ -1,5 +1,7 @@
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
+const User = require('../models/user');
+const jwt = require('jsonwebtoken');
 
 blogsRouter.get('/', (request, response) => {
   Blog.find({})
@@ -9,18 +11,32 @@ blogsRouter.get('/', (request, response) => {
     });
 });
 
-blogsRouter.post('/', (request, response, next) => {
-  const blog = new Blog(request.body);
+blogsRouter.post('/', async (request, response) => {
+  const body = request.body;
 
-  blog
-    .save()
-    .then((result) => {
-      response.status(201).json(result);
-    })
-    .catch((error) => next(error));
+  const user = await User.findById(request.user);
+
+  const blog = new Blog({
+    title: body.title,
+    url: body.url,
+    likes: body.likes,
+    author: body.author,
+    user: user._id,
+  });
+
+  const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
+  response.status(201).json(savedBlog);
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
+  if (blog.user.toString() !== request.user.toString()) {
+    return response.status(403).json({ error: 'access forbidden' });
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
   response.status(204).end();
 });
